+++ mullvad-cli/src/cmds/log.rs
@@ -6,12 +6,27 @@ use mullvad_management_interface::MullvadProxyClient;
 #[derive(Subcommand, Debug)]
 pub enum Log {
     /// Set the log level for the daemon
-    SetLevel {
-        /// The log level to set
-        level: String,
-    },
+    SetLevel { level: Level },
     /// Follow live updates to the daemon log file. Analogue to running `tail -f` on the daemon log file.
     Listen,
+}
+
+/// See <https://docs.rs/log/latest/log/enum.Level.html>
+#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq, clap::ValueEnum)]
+#[repr(usize)]
+pub enum Level {
+    /// Turn off logging.
+    Off = 0,
+    /// Log very serious errors.
+    Error = 1,
+    /// Log hazardous situations.
+    Warn = 2,
+    /// Log useful information.
+    Info = 3,
+    /// Log lower priority information.
+    Debug = 4,
+    /// Log very low priority, often extremely verbose, information.
+    Trace = 5,
 }
 
 impl Log {
@@ -37,8 +52,9 @@ async fn on_listen() -> std::result::Result<(), anyhow::Error> {
     Ok(())
 }
 
-async fn set_level(level: String) -> std::result::Result<(), anyhow::Error> {
+async fn set_level(level: Level) -> std::result::Result<(), anyhow::Error> {
     let mut rpc = MullvadProxyClient::new().await?;
+    let level = (level as usize).to_string();
     rpc.set_log_filter(level).await?;
     Ok(())
 }